+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog - v1.0.6
 - Show changelog if new update was installed
 - Added Link to full changelog to about page
+- Improved update error notification
 
 
 ## Changelog - v1.0.5